+++ app/src/main/java/com/fibelatti/photowidget/home/UserPreferences.kt
@@ -1,6 +1,0 @@
-package com.fibelatti.photowidget.home
-
-data class UserPreferences(
-    val appearance: Appearance,
-    val dynamicColors: Boolean,
-)