+++ app/src/main/java/com/fibelatti/photowidget/model/PhotoWidgetTapAction.kt
@@ -4,12 +4,12 @@ import androidx.annotation.StringRes
 import com.fibelatti.photowidget.R
 
 enum class PhotoWidgetTapAction(
-    @StringRes val title: Int,
+    @StringRes val label: Int,
 ) {
 
-    NONE(title = R.string.photo_widget_configure_tap_action_none),
-    VIEW_FULL_SCREEN(title = R.string.photo_widget_configure_tap_action_view_full_screen),
-    VIEW_NEXT_PHOTO(title = R.string.photo_widget_configure_tap_action_next_photo),
-    APP_SHORTCUT(title = R.string.photo_widget_configure_tap_action_app_shortcut),
+    NONE(label = R.string.photo_widget_configure_tap_action_none),
+    VIEW_FULL_SCREEN(label = R.string.photo_widget_configure_tap_action_view_full_screen),
+    VIEW_NEXT_PHOTO(label = R.string.photo_widget_configure_tap_action_next_photo),
+    APP_SHORTCUT(label = R.string.photo_widget_configure_tap_action_app_shortcut),
     ;
 }