+++ app/src/main/java/com/fibelatti/photowidget/preferences/Appearance.kt
@@ -1,4 +1,4 @@
-package com.fibelatti.photowidget.home
+package com.fibelatti.photowidget.preferences
 
 enum class Appearance {
     FOLLOW_SYSTEM,